feat: 필드 기본 타입으로 변경

+++ backend/PollockHub/src/main/java/com/pollock/pollockhub/engine/dto/request/EngineAnalysisRequestDTO.java
@@ -11,6 +11,6 @@ public class EngineAnalysisRequestDTO {
     private String channelKey;
     private EngineType engineType;
     private String fen;
-    private Integer multipv;
-    private Long movetime;
+    private int multipv;
+    private int movetime;
 }
